Refactor scripts.js and styles.js
