+++ lib/createProductionTask.js
@@ -1,0 +1,8 @@
+const productionDefaults = {
+  minify: true,
+  sourcemaps: false,
+};
+
+module.exports = (task) => {
+  return () => task(productionDefaults);
+}
